replace: long => int32_t add struct IPOINT{int32_t x, y;} replace POINT with IPOINT, when it was necessary

+++ src/apps/engine/src/core_impl.h
@@ -74,7 +74,7 @@ class CoreImpl : public Core
     void AppState(bool state);
     uint32_t MakeHashValue(const char *string);
     VMA *FindVMA(const char *class_name);
-    VMA *FindVMA(long hash);
+    VMA *FindVMA(int32_t hash);
     //------------------------------------------------------------------------------------------------
     // API functions : (virtual API)
 
@@ -117,16 +117,16 @@ class CoreImpl : public Core
     // return current fps
     uint32_t EngineFps() override;
     // set fixed delta time mode, (-1) - off
-    void SetDeltaTime(long delta_time) override;
+    void SetDeltaTime(int32_t delta_time) override;
     uint32_t GetDeltaTime() override;
     uint32_t GetRDeltaTime() override;
     //
     VDATA *Event(const char *Event_name, const char *Format = nullptr, ...) override;
     uint32_t PostEvent(const char *Event_name, uint32_t post_time, const char *Format, ...) override;
 
-    void *GetSaveData(const char *file_name, long &data_size) override;
+    void *GetSaveData(const char *file_name, int32_t &data_size) override;
 
-    bool SetSaveData(const char *file_name, void *data_ptr, long data_size) override;
+    bool SetSaveData(const char *file_name, void *data_ptr, int32_t data_size) override;
 
     uint32_t SetScriptFunction(IFUNCINFO *pFuncInfo) override;
