+++ src/apps/engine/src/token.cpp
@@ -642,7 +642,7 @@ S_TOKEN_TYPE TOKEN::FormatGet()
 }
 
 // copy argument data to buffer and close the termination 0
-long TOKEN::SetTokenData(const char *pointer, bool bKeepControlSymbols)
+int32_t TOKEN::SetTokenData(const char *pointer, bool bKeepControlSymbols)
 {
     // if(!IsOperator(pointer,Data_size))
     const auto Data_size = StopArgument(pointer, bKeepControlSymbols);
@@ -687,9 +687,9 @@ ptrdiff_t TOKEN::SetNTokenData(const char *pointer, ptrdiff_t Data_size)
 // search throw the program code until find non significant argument character:
 // SPACE,TAB,0,'\r','\n'
 // return number of significant symbols
-long TOKEN::StopArgument(const char *pointer, bool bKeepControlSymbols)
+int32_t TOKEN::StopArgument(const char *pointer, bool bKeepControlSymbols)
 {
-    long size = 0;
+    int32_t size = 0;
     auto bDot = false;
     auto bOnlyDigit = true;
     do
@@ -986,7 +986,7 @@ bool TOKEN::IsOperator(char * pointer)
 }
 */
 /*
-bool TOKEN::IsOperator(char * pointer, long & syms)
+bool TOKEN::IsOperator(char * pointer, int32_t & syms)
 {
     DWORD operators_num,n,i;
     bool found;
@@ -1046,8 +1046,8 @@ bool TOKEN::StepBack()
 S_TOKEN_TYPE TOKEN::ProcessToken(char *&pointer, bool bKeepData)
 {
     char sym;
-    // long keywords_num;
-    // long n;
+    // int32_t keywords_num;
+    // int32_t n;
     char *pBase;
 
     pointer += SetTokenData(pointer, bKeepData);
@@ -1195,7 +1195,7 @@ S_TOKEN_TYPE TOKEN::ProcessToken(char *&pointer, bool bKeepData)
     return eTokenType;
 }
 
-long TOKEN::TokenLines()
+int32_t TOKEN::TokenLines()
 {
     return Lines_in_token;
 }
