+++ src/apps/engine/src/token.h
@@ -161,10 +161,10 @@ class TOKEN
     THLINE KeywordsHash[TOKENHASHTABLE_SIZE];
     S_TOKEN_TYPE eTokenType;
     ptrdiff_t TokenDataBufferSize;
-    long Lines_in_token;
+    int32_t Lines_in_token;
     char *pTokenData;
     ptrdiff_t ProgramSteps[PROGRAM_STEPS_CACHE];
-    long ProgramStepsNum;
+    int32_t ProgramStepsNum;
     char *Program;
     char *ProgramBase;
     uint32_t dwKeywordsNum;
@@ -189,9 +189,9 @@ class TOKEN
     S_TOKEN_TYPE GetType();
     void CacheToken(const char *pointer);
     bool StepBack();
-    long SetTokenData(const char *pointer, bool bKeepControlSymbols = false);
+    int32_t SetTokenData(const char *pointer, bool bKeepControlSymbols = false);
     ptrdiff_t SetNTokenData(const char *pointer, ptrdiff_t Data_size);
-    long StopArgument(const char *pointer, bool bKeepControlSymbols = false);
+    int32_t StopArgument(const char *pointer, bool bKeepControlSymbols = false);
     void StartArgument(char *&pointer, bool bKeepControlSymbols = false);
     const char *GetTypeName(S_TOKEN_TYPE code);
     const char *GetTypeName();
@@ -199,7 +199,7 @@ class TOKEN
     bool Is(S_TOKEN_TYPE ttype);
     bool IsNumber(const char *pointer);
     bool IsFloatNumber(const char *pointer);
-    long TokenLines();
+    int32_t TokenLines();
     void LowCase();
 
     S_TOKEN_TYPE FormatGet();
