+++ src/libs/animals/src/t_seagulls.cpp
@@ -70,7 +70,7 @@ void TSeagulls::Init()
 }
 
 //--------------------------------------------------------------------
-uint64_t TSeagulls::ProcessMessage(long _code, MESSAGE &message)
+uint64_t TSeagulls::ProcessMessage(int32_t _code, MESSAGE &message)
 {
     const uint32_t outValue = 0;
 
@@ -146,7 +146,7 @@ void TSeagulls::Execute(uint32_t _dTime)
             const auto oldR = seagulls[i].radius;
             seagulls[i].radius = SEAGULL_MIN_RADIUS + rand(maxRadius);
             seagulls[i].circleTimePassed = 0;
-            seagulls[i].circleTime = static_cast<long>(rand(static_cast<float>(maxCircleTime)));
+            seagulls[i].circleTime = static_cast<int32_t>(rand(static_cast<float>(maxCircleTime)));
             if ((seagulls[i].circleTime) < (maxCircleTime / 20))
                 seagulls[i].circleTime = maxCircleTime / 20;
             const auto sinA = sinf(seagulls[i].a);
