+++ src/libs/animation/src/animation_info.cpp
@@ -32,7 +32,7 @@ AnimationInfo::~AnimationInfo()
 }
 
 // Create bones
-void AnimationInfo::CreateBones(long numbones)
+void AnimationInfo::CreateBones(int32_t numbones)
 {
     Assert(bone == nullptr || numBones == 0);
     Assert(numbones > 0 && numbones <= 256);
@@ -41,7 +41,7 @@ void AnimationInfo::CreateBones(long numbones)
 }
 
 // Create action
-ActionInfo *AnimationInfo::AddAction(const char *anctionName, long startframe, long endframe)
+ActionInfo *AnimationInfo::AddAction(const char *anctionName, int32_t startframe, int32_t endframe)
 {
     Assert(anctionName);
     // Looking for repetition
